Refactor conversion code for pypi to be in shared utilities
